vault_analytics/requirements.txt
pywebview>=5.0
fastapi>=0.100.0
uvicorn>=0.23.0
